+++ libs/core/src/system_state/ui_colors.ts
@@ -14,42 +14,23 @@ export class UIColors {
     return newOnEvent(cb, this, SeelenEvent.ColorsChanged);
   }
 
-  static default(): UIColors {
-    return new this({
-      background: "#ffffff",
-      foreground: "#000000",
-      accent_darkest: "#990000",
-      accent_darker: "#aa0000",
-      accent_dark: "#bb0000",
-      accent: "#cc0000",
-      accent_light: "#dd0000",
-      accent_lighter: "#ee0000",
-      accent_lightest: "#ff0000",
-      complement: null,
-    });
-  }
-
   setAsCssVariables(): void {
     const oldStyles = new RuntimeStyleSheet("@deprecated/system-colors");
     const newStyles = new RuntimeStyleSheet("@runtime/system-colors");
 
-    for (const [key, value] of Object.entries(this.inner)) {
-      if (typeof value !== "string") {
-        continue;
-      }
-
-      const color = Color.fromHex(value);
-      const { r, g, b } = color;
+    for (const [key, _color] of Object.entries(this.inner)) {
+      const color = new Color(_color);
+      const { r, g, b } = _color;
 
       // replace rust snake case with kebab case
       const name = key.replace("_", "-");
 
       // @deprecated old names
-      oldStyles.addVariable(`--config-${name}-color`, value.slice(0, 7));
+      oldStyles.addVariable(`--config-${name}-color`, color.toHexString());
       oldStyles.addVariable(`--config-${name}-color-rgb`, `${r}, ${g}, ${b}`);
 
       if (name.startsWith("accent")) {
-        newStyles.addVariable(`--system-${name}-color`, value.slice(0, 7));
+        newStyles.addVariable(`--system-${name}-color`, color.toHexString());
       }
     }