+++ src/ui/react/fancy-toolbar/app.tsx
@@ -1,6 +1,3 @@
-import { $system_colors } from "libs/ui/react/utils/signals.ts";
-import { useDarkMode } from "libs/ui/react/utils/styling.ts";
-import { ConfigProvider, theme } from "antd";
 import { Widget } from "@seelen-ui/lib";
 import { ErrorBoundary } from "../weg/components/Error/index.tsx";
 import { ErrorFallback } from "./components/Error/index.tsx";
@@ -10,8 +7,6 @@ import { $lastFocusedOnMonitor } from "./modules/shared/state/windows.ts";
 import { useEffect } from "preact/hooks";
 
 export function App() {
-  const isDarkMode = useDarkMode();
-
   useEffect(() => {
     Widget.self.ready({ show: false });
   }, []);
@@ -26,24 +21,8 @@ export function App() {
   });
 
   return (
-    <ConfigProvider
-      theme={{
-        token: {
-          colorPrimary: isDarkMode ? $system_colors.value.accent_light : $system_colors.value.accent_dark,
-        },
-        components: {
-          Calendar: {
-            fullBg: "transparent",
-            fullPanelBg: "transparent",
-            itemActiveBg: "transparent",
-          },
-        },
-        algorithm: isDarkMode ? theme.darkAlgorithm : theme.defaultAlgorithm,
-      }}
-    >
-      <ErrorBoundary fallback={<ErrorFallback />}>
-        <FancyToolbar />
-      </ErrorBoundary>
-    </ConfigProvider>
+    <ErrorBoundary fallback={<ErrorFallback />}>
+      <FancyToolbar />
+    </ErrorBoundary>
   );
 }