+++ src/ui/react/settings/app.tsx
@@ -1,7 +1,7 @@
 import { useDarkMode } from "libs/ui/react/utils/styling.ts";
 import { ConfigProvider, theme } from "antd";
 import { useEffect } from "react";
-import { Widget } from "@seelen-ui/lib";
+import { Color, Widget } from "@seelen-ui/lib";
 
 import { Routing } from "./router.tsx";
 import { ThumbnailGeneratorModal } from "./components/ThumbnailGeneratorModal/index.tsx";
@@ -21,12 +21,14 @@ export function App() {
     Widget.self.ready();
   }, []);
 
+  const color = new Color(isDarkMode ? uiColors.value.accent_light : uiColors.value.accent_dark);
+
   return (
     <ConfigProvider
       componentSize="small"
       theme={{
         token: {
-          colorPrimary: isDarkMode ? uiColors.value.accent_light : uiColors.value.accent_dark,
+          colorPrimary: color.toHexString(),
         },
         algorithm: isDarkMode ? theme.darkAlgorithm : theme.defaultAlgorithm,
       }}