+++ src/ui/react/weg/app.tsx
@@ -1,7 +1,3 @@
-import { $system_colors } from "libs/ui/react/utils/signals.ts";
-import { useDarkMode } from "libs/ui/react/utils/styling.ts";
-import { ConfigProvider, theme } from "antd";
-
 import { ErrorBoundary } from "./components/Error/index.tsx";
 import { SeelenWeg } from "./modules/bar/index.tsx";
 import { useSignalEffect } from "@preact/signals";
@@ -9,8 +5,6 @@ import { $lastFocusedOnMonitor } from "./modules/shared/state/windows.ts";
 import { Widget } from "@seelen-ui/lib";
 
 export function App() {
-  const isDarkMode = useDarkMode();
-
   useSignalEffect(() => {
     const fullscreened = !!$lastFocusedOnMonitor.value?.isFullscreened;
     if (fullscreened) {
@@ -21,18 +15,8 @@ export function App() {
   });
 
   return (
-    <ConfigProvider
-      componentSize="small"
-      theme={{
-        token: {
-          colorPrimary: isDarkMode ? $system_colors.value.accent_light : $system_colors.value.accent_dark,
-        },
-        algorithm: isDarkMode ? theme.darkAlgorithm : theme.defaultAlgorithm,
-      }}
-    >
-      <ErrorBoundary fallback={<div>Something went wrong</div>}>
-        <SeelenWeg />
-      </ErrorBoundary>
-    </ConfigProvider>
+    <ErrorBoundary fallback={<div>Something went wrong</div>}>
+      <SeelenWeg />
+    </ErrorBoundary>
   );
 }