+++ changelog.md
@@ -7,6 +7,7 @@
 - Support for game mode.
 - Close webviews while in game mode.
 - Add new unstable optimizations flag under 'for developers' tab.
+- dynamic accent color by wallpaper.
 
 ### enhancements
 
@@ -16,6 +17,7 @@
 
 - multiline tooltips.
 - show desktop not working.
+- flicker on wallpapers on change.
 
 ## [2.5.9]